rpcs3: Move input to its own directory
