+++ rpcs3/Emu/Io/PadHandler.cpp
@@ -1,6 +1,6 @@
 ﻿#include "stdafx.h"
 #include "PadHandler.h"
-#include "pad_thread.h"
+#include "Input/pad_thread.h"
 
 cfg_input g_cfg_input;
