+++ rpcs3/Input/ds3_pad_handler.h
