+++ rpcs3/Input/ds4_pad_handler.cpp
