+++ rpcs3/Input/keyboard_pad_handler.h
