+++ rpcs3/Input/xinput_pad_handler.h
