+++ rpcs3/rpcs3.vcxproj
@@ -331,17 +331,17 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="basic_keyboard_handler.cpp" />
-    <ClCompile Include="basic_mouse_handler.cpp" />
     <ClCompile Include="display_sleep_control.cpp" />
-    <ClCompile Include="ds3_pad_handler.cpp" />
-    <ClCompile Include="ds4_pad_handler.cpp" />
-    <ClCompile Include="evdev_joystick_handler.cpp" />
-    <ClCompile Include="keyboard_pad_handler.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="main_application.cpp" />
-    <ClCompile Include="mm_joystick_handler.cpp" />
-    <ClCompile Include="pad_thread.cpp" />
+    <ClCompile Include="Input\basic_keyboard_handler.cpp" />
+    <ClCompile Include="Input\basic_mouse_handler.cpp" />
+    <ClCompile Include="Input\ds3_pad_handler.cpp" />
+    <ClCompile Include="Input\ds4_pad_handler.cpp" />
+    <ClCompile Include="Input\evdev_joystick_handler.cpp" />
+    <ClCompile Include="Input\keyboard_pad_handler.cpp" />
+    <ClCompile Include="Input\mm_joystick_handler.cpp" />
+    <ClCompile Include="Input\pad_thread.cpp" />
     <ClCompile Include="QTGeneratedFiles\Debug - LLVM\moc_about_dialog.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release - LLVM|x64'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</ExcludedFromBuild>
@@ -1193,7 +1193,7 @@
     <ClCompile Include="rpcs3qt\save_data_dialog.cpp" />
     <ClCompile Include="rpcs3qt\save_data_list_dialog.cpp" />
     <ClCompile Include="rpcs3qt\settings_dialog.cpp" />
-    <ClCompile Include="xinput_pad_handler.cpp" />
+    <ClCompile Include="Input\xinput_pad_handler.cpp" />
     <ClCompile Include="\rpcs3qt\*.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -1215,7 +1215,7 @@
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='Release|x64'">$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
     </CustomBuild>
-    <CustomBuild Include="basic_keyboard_handler.h">
+    <CustomBuild Include="Input\basic_keyboard_handler.h">
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Moc%27ing basic_keyboard_handler.h...</Message>
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">Moc%27ing basic_keyboard_handler.h...</Message>
       <Outputs Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">.\QTGeneratedFiles\$(ConfigurationName)\moc_%(Filename).cpp</Outputs>
@@ -1233,7 +1233,7 @@
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='Release|x64'">$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
     </CustomBuild>
-    <CustomBuild Include="basic_mouse_handler.h">
+    <CustomBuild Include="Input\basic_mouse_handler.h">
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Moc%27ing basic_mouse_handler.h...</Message>
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">Moc%27ing basic_mouse_handler.h...</Message>
       <Outputs Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">.\QTGeneratedFiles\$(ConfigurationName)\moc_%(Filename).cpp</Outputs>
@@ -1252,10 +1252,10 @@
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">$(QTDIR)\bin\moc.exe;%(FullPath);$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
     </CustomBuild>
     <ClInclude Include="display_sleep_control.h" />
-    <ClInclude Include="ds3_pad_handler.h" />
-    <ClInclude Include="ds4_pad_handler.h" />
-    <ClInclude Include="evdev_joystick_handler.h" />
-    <ClInclude Include="keyboard_pad_handler.h" />
+    <ClInclude Include="Input\ds3_pad_handler.h" />
+    <ClInclude Include="Input\ds4_pad_handler.h" />
+    <ClInclude Include="Input\evdev_joystick_handler.h" />
+    <ClInclude Include="Input\keyboard_pad_handler.h" />
     <CustomBuild Include="rpcs3qt\gs_frame.h">
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Moc%27ing gs_frame.h...</Message>
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">Moc%27ing gs_frame.h...</Message>
@@ -1535,7 +1535,7 @@
       <Outputs Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">.\QTGeneratedFiles\$(ConfigurationName)\moc_%(Filename).cpp</Outputs>
       <Command Condition="'$(Configuration)|$(Platform)'=='Debug - LLVM|x64'">"$(QTDIR)\bin\moc.exe"  "%(FullPath)" -o ".\QTGeneratedFiles\$(ConfigurationName)\moc_%(Filename).cpp"  -D_WINDOWS -DUNICODE -DWIN32 -DWIN64 -DQT_OPENGL_LIB -DQT_WIDGETS_LIB -DQT_GUI_LIB -DQT_QML_LIB -DQT_NETWORK_LIB -DQT_CORE_LIB -DQT_WINEXTRAS_LIB -DQT_CONCURRENT_LIB -D%(PreprocessorDefinitions)  "-I.\..\3rdparty\libusb\libusb" "-I$(VULKAN_SDK)\Include" "-I.\..\3rdparty\minidx12\Include" "-I$(QTDIR)\include" "-I$(QTDIR)\include\QtOpenGL" "-I$(QTDIR)\include\QtWidgets" "-I$(QTDIR)\include\QtGui" "-I$(QTDIR)\include\QtANGLE" "-I$(QTDIR)\include\QtQml" "-I$(QTDIR)\include\QtNetwork" "-I$(QTDIR)\include\QtCore" "-I.\debug" "-I$(QTDIR)\mkspecs\win32-msvc2015" "-I.\QTGeneratedFiles\$(ConfigurationName)" "-I.\QTGeneratedFiles" "-I$(QTDIR)\include\QtWinExtras" "-I$(QTDIR)\include\QtConcurrent"</Command>
     </CustomBuild>
-    <ClInclude Include="pad_thread.h" />
+    <ClInclude Include="Input\pad_thread.h" />
     <ClInclude Include="QTGeneratedFiles\ui_about_dialog.h" />
     <ClInclude Include="QTGeneratedFiles\ui_main_window.h" />
     <ClInclude Include="QTGeneratedFiles\ui_pad_settings_dialog.h" />
@@ -2007,7 +2007,7 @@
     <ClInclude Include="rpcs3qt\trophy_notification_frame.h" />
     <ClInclude Include="rpcs3qt\trophy_notification_helper.h" />
     <ClInclude Include="rpcs3qt\user_account.h" />
-    <ClInclude Include="xinput_pad_handler.h" />
+    <ClInclude Include="Input\xinput_pad_handler.h" />
     <ClInclude Include="\rpcs3qt\*.h" />
   </ItemGroup>
   <ItemGroup>
