+++ rpcs3/rpcs3qt/pad_settings_dialog.cpp
@@ -13,15 +13,15 @@
 
 #include "Emu/Io/Null/NullPadHandler.h"
 
-#include "keyboard_pad_handler.h"
-#include "ds3_pad_handler.h"
-#include "ds4_pad_handler.h"
+#include "Input/keyboard_pad_handler.h"
+#include "Input/ds3_pad_handler.h"
+#include "Input/ds4_pad_handler.h"
 #ifdef _WIN32
-#include "xinput_pad_handler.h"
-#include "mm_joystick_handler.h"
+#include "Input/xinput_pad_handler.h"
+#include "Input/mm_joystick_handler.h"
 #endif
 #ifdef HAVE_LIBEVDEV
-#include "evdev_joystick_handler.h"
+#include "Input/evdev_joystick_handler.h"
 #endif
 
 inline std::string sstr(const QString& _in) { return _in.toStdString(); }
